# Docs: Add render performance measurement in test environments