--- a/src/MudBlazor.Docs/Components/DocsPage.razor
+++ b/src/MudBlazor.Docs/Components/DocsPage.razor
@@ -1,7 +1,7 @@
 ﻿<MudMainContent Class="mudblazor-main-content">
     <MudContainer MaxWidth="@MaxWidth">
         <div id="docspage" class="docs-page">
-            <CascadingValue Value="this" IsFixed="true">
+            <CascadingValue Value="this" IsFixed>
                 @ChildContent
             </CascadingValue>
         </div>