We have a way to measure docs page render times now:

This is the base line, showing WASM render times to be way too high, BSS times are acceptable:

```
BSS:
components/alert rendered in 80,573ms
components/appbar rendered in 58,1864ms
components/avatar rendered in 82,9861ms
components/badge rendered in 132,3375ms
components/breadcrumbs rendered in 52,4078ms

WASM:
components/alert rendered in 600ms
components/appbar rendered in 496ms
components/avatar rendered in 546ms
components/badge rendered in 571ms
components/breadcrumbs rendered in 515ms
```

Temporarily the measurement is shown right on the top of the page.
![image](https://user-images.githubusercontent.com/44090/141652315-cd72725e-a355-4c3a-971b-e5ac201619e0.png)

Note, to get a true measurement you need to do a warm-up. You need to click a few times between Alert and AppBar to get a representative measurement.

So far I found this: (the following measurements focus on WASM and only on the alert page)

```
baseline: ~600ms
w/o page nav menu: ~514ms (so the Menu on the right costs about 80ms)
w/o footer: ~495ms (footer costs about 20ms)
w/o any content: 11ms (the page content costs ~500ms)
```

I'll deal with menu and footer later, let's first get the page content loading times down

```
new baseline without nav and footer: ~500ms 
w/o API link: ~216ms (API link costs about 300ms)
w/o SEO tags: ~200ms (cheap)
w/o example source code: ~80ms (gains 120ms for alert)
```