@henon 2021-11-13
I want to keep it in dev.mudblazor.com also, so I found a way to display it beneath the heading and description where it probably won't irritate anybody (hopefully). 

![image](https://user-images.githubusercontent.com/44090/141657340-9bbd4ebb-0aa2-422f-a849-5e3ef2fa5b3b.png)