@henon 2021-11-13
yeah, it is an idea. we have it logging to console in all environments. the measurement is only displayed on localhost or dev.mudblazor.com. I'll leave dealing with it visually to Jonny if he wants to. 

I'll merge this so others can try different strategies. Mike wants to try pre-rendering. I'll start chipping away at some of the most obvious problems like the API link and the in-page menu in separate PRs