@henon 2021-11-13
The change is online now: https://dev.mudblazor.com/components/alert