@henon 2021-11-13
very interesting is that on dev.mudblazor.com the WASM time is half the time on my machine with DEBUG and RELEASE!

Alert local DEBUG: 600ms
Alert local RELEASE: 600ms
Alert dev.mudblazor.com: 300ms

@mikes-gh can you explain this?