@henon 2021-11-13
and don't say my machine is slow >:(