@iwis 2021-12-07
Can you also make it working for the API documentation, please? It doesn't refresh there. I had to go to some page in the "Components" menu, then go back to API documentation page to refresh it - but it is tedious and I don't know if the value is correct.

https://user-images.githubusercontent.com/8080496/145002136-132f937c-a118-4c53-8eb4-616d82dc1af7.mp4