@henon 2021-12-07
Sure, I know this bug. Will look at it today