@iwis 2021-12-07
I have one more issue. I was switching between Bar chart and Line chart pages: Bar chart page, Line chart page, Bar chart page, Line chart page, Bar chart page. Both pages have no sections so the page is displayed only after the whole content is ready to be shown. Bar chart page is opened in about 100 ms according to this timer. But Developer Tools shows that it is rather 300 ms. I also perceive it to be rather 300 ms than 100 ms. So why it says "Rendered in 103 ms" - what does this timer measure?

![image](https://user-images.githubusercontent.com/8080496/145008954-d604912c-1f41-4f22-b956-16d84bb516c9.png)