@henon 2021-12-07
It should measure the time it takes to render the first three sections. Maybe this is caused by the same bug. Try again after I fixed it for api pages. Of course, using developer tools to trace performance is most reliable. This stopwatch is just an approximation. It measures the time from before rendering the page until the timer item's OnAfterRender method is called.

It was very useful for me to get page load times down significantly.