@henon 2021-12-07
ok. should be fixed now. 

by the way, the measurement does not measure the whole page rendering any more. since my optimization in later PRs, where I implemented incremental rendering for more reactiveness, it measures only the initial render which renders only a part of the page.