@henon 2021-12-07
To be exact, the timer shows the time it takes until the user gets to see the first screen of the page (even if lower parts of it are still loading).