@iwis 2021-12-07
On Wasm it measures exactly the time between mouse clicks now :D