@iwis 2021-12-07
On Wasm it measures exactly the time between mouse clicks now :D i.e. in the https://dev.mudblazor.com/wasm/api/ menu :D