@iwis 2021-12-07
Yes, now it shows clearly correlation with the speed of pages in API menu. For fast loading pages it shows 50-100 ms. For slow pages (with many properties) it shows 200-275 ms. I will use it tomorrow to check if I can speed up slow pages. Thanks!